Split synchronization helpers in CrowbarPacemakerSynchronization module
